http client spi support by shared context
